requirements.txt
# Core
Django
django-environ
gunicorn
django-jazzmin
# psycopg2

# For the persistence stores
psycopg2-binary

# Rest apis
djangorestframework
djangorestframework_simplejwt
django-cors-headers
drf-yasg
whitenoise
shortuuid

# Code quality
# ------------------------------------------------------------------------------
flake8==4.0.1
flake8-isort==4.1.1
coverage==6.4.2
black==22.6.0
pylint-django==2.5.3
pre-commit==2.20.0

# Certificates
cryptography~=37.0.4
